requirements.txt
matplotlib==3.1.2
numpy==1.17.4
Pillow==9.5.0
prenlp==0.0.13
scikit_learn==1.2.2
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tqdm==4.62.3
